Replace all existing commands with brigadier commands.
- Remove plugin.yml in favor of paper-plugin.yml
- Switch to a Permissions enum for checking permissions, and or supplying the information to command literals.
- Switch to using doubles for prices instead of longs, as Vault internally uses doubles anyway.
- Other misc changes like not having to set the configuration section to the data.yml
- Remove unused methods because I updated the instances where they were used to be better.

+++ api/src/main/java/us/crazycrew/api/storage/IStorageHolder.java
@@ -16,7 +16,7 @@ public abstract class IStorageHolder {
             @NonNull final UUID uuid,
             @NonNull final String name,
             @NonNull final String base64,
-            final long price,
+            final double price,
             @NonNull final ShopType shopType
     );
