+++ paper/src/main/java/com/badbones69/crazyauctions/api/events/AuctionBuyEvent.java
@@ -18,7 +18,7 @@ public class AuctionBuyEvent extends Event {
     private static final HandlerList handlers = new HandlerList();
 
     private final Player player;
-    private final long price;
+    private final double price;
     private final ItemStack item;
     
     /**
@@ -27,7 +27,7 @@ public class AuctionBuyEvent extends Event {
      * @param item The item that was bought.
      * @param price The price of the item.
      */
-    public AuctionBuyEvent(Player player, ItemStack item, long price) {
+    public AuctionBuyEvent(Player player, ItemStack item, double price) {
         this.player = player;
         this.item = item;
         this.price = price;
@@ -49,7 +49,7 @@ public class AuctionBuyEvent extends Event {
         return this.item;
     }
     
-    public long getPrice() {
+    public double getPrice() {
         return this.price;
     }
 }
