+++ paper/src/main/java/com/badbones69/crazyauctions/api/events/AuctionListEvent.java
@@ -8,28 +8,25 @@ import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.NotNull;
 
 /**
- *
  * @author BadBones69
  *
  * This event is fired when a new item is listed onto the auction house.
- *
  */
 public class AuctionListEvent extends Event {
     
     private static final HandlerList handlers = new HandlerList();
     private final Player player;
-    private final long price;
+    private final double price;
     private final ShopType shop;
     private final ItemStack item;
-    
+
     /**
-     *
      * @param player The player who is listing the item.
      * @param shop The shop type the item is being listed to.
      * @param item The item being listed.
      * @param price The price the item is being listed for.
      */
-    public AuctionListEvent(Player player, ShopType shop, ItemStack item, long price) {
+    public AuctionListEvent(Player player, ShopType shop, ItemStack item, double price) {
         this.player = player;
         this.shop = shop;
         this.item = item;
@@ -55,8 +52,8 @@ public class AuctionListEvent extends Event {
     public ItemStack getItem() {
         return this.item;
     }
-    
-    public long getPrice() {
+
+    public double getPrice() {
         return this.price;
     }
 }
