+++ paper/src/main/java/com/badbones69/crazyauctions/api/events/AuctionWinBidEvent.java
@@ -17,7 +17,7 @@ public class AuctionWinBidEvent extends Event {
     
     private static final HandlerList handlers = new HandlerList();
     private final Player player;
-    private final long bid;
+    private final double bid;
     private final ItemStack item;
     
     /**
@@ -26,7 +26,7 @@ public class AuctionWinBidEvent extends Event {
      * @param item The item that was won.
      * @param bid The bid that was placed on the item.
      */
-    public AuctionWinBidEvent(Player player, ItemStack item, long bid) {
+    public AuctionWinBidEvent(Player player, ItemStack item, double bid) {
         this.player = player;
         this.item = item;
         this.bid = bid;
@@ -48,7 +48,7 @@ public class AuctionWinBidEvent extends Event {
         return this.item;
     }
     
-    public long getBid() {
+    public double getBid() {
         return this.bid;
     }
 }
